requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==1.1.0
gspread==5.12.0
google-auth==2.25.2
gunicorn==21.2.0
pandas
anthropic
